Revert "世界坐标"

+++ scenes/space_rock/SpaceRock.cs
@@ -34,7 +34,7 @@ public partial class SpaceRock : RigidBody2D,IAsteroid
 		// 初始化AsteroidData
 		AsteroidData = new AsteroidData
 		{
-			Health = (int)hp,
+			Health = hp,
 			Loot = dropId
 		};
 		
@@ -72,7 +72,7 @@ public partial class SpaceRock : RigidBody2D,IAsteroid
 					if (animatedSprite != null)
 					{
 						animatedSprite.Visible = false;
-						ShootArea.QueueFree();
+                        ShootArea.QueueFree();
 						Timer tempTimer = new Timer();
 						tempTimer.WaitTime = 2.0f; // 粒子播放时长
 						tempTimer.Timeout += () => QueueFree();
